pyproject.toml
[project]
name = "bsod-analyzer"
version = "1.0.0"
description = "Windows BSOD Dump Analysis Tool with AI"
authors = [{name = "Your Name"}]
license = {text = "MIT"}
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    "minidump>=0.0.24",
    "kdmp-parser>=0.7.0",
    "click>=8.1.0",
    "rich>=14.0.0",
    "zhipuai>=2.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
bsod = "bsod_analyzer.cli.main:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]
